requirements.txt
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.1
Django>=5.0.7
django-debug-toolbar>=4.2.0
django-crispy-forms>=1.14.0
crispy-bootstrap5>=0.7
psycopg2-binary>=2.9.5
pytest-django>=4.8.0
redis>=5.0.1
ipython>=8.0.1
django-environ>=0.11.2
argon2-cffi>=21.3.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
drf-yasg>=1.1.0rc8